+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.3.2 (not yet published)
+
+- Upgrade dependencies
+- Fix waiting for animation: Make sure the animation happens on the actual modal element, and not a child.
+
+## 2.3.1
+
+- Add `contenteditable="true"` support for focus trap
+
 ## 2.3.0
 
 - Upgrade dependencies